mkdir / modified readme
