+++ README.md
@@ -1,5 +1,34 @@
-# Cosadama
- __Cosadama 코사다마__
+# COSADAMA
 
-Cosadama is the programming boot camp by sociology students of Chung-Ang university, Seoul, Korea. The group name "Cosadama" comes from the compound of 
-"CODING", __"사회학(sociology)"__ which sounds 'sa-hui-hak' and **"好事多魔"** which is a proverb meaning that you need to hard work to go right way and sounds 'ho-sa-da-ma'. The name means that we all finally go right way(maybe could be top programmer) after taking all the struggles, such as endless errors, back-ache etc. Then, Look forward to what we are! Love you all from sociology dept. 
+
+
+![cosadama](./COSADAMA_LOGO(300).png)
+
+## __Cosadama 코사다마__
+
+*The best way to do something is to do with someone to produce something.*
+
+Cosadama is the programming boot camp by sociology students of Chung-Ang university, Seoul, Korea. The group name "Cosadama" comes from the compound of  "CODING", __"사회학(sociology)"__ which sounds 'sa-hui-hak' and **"好事多魔"** which is a proverb meaning that you need to hard work to go right way and sounds 'ho-sa-da-ma'. The name means that we all finally go right way(maybe could be top programmer) after taking all the struggles, such as endless errors, back-ache etc. Then, Look forward to what we are! Love you all from sociology dept. 
+
+## __Audience__
+
+Our target audience for the COSADAMA is : 
+
+Beginner of Beginner, a person who has just begun the python or says 'Python is just all greek to me'.
+
+## __Objectives__
+
+* Knowledge of python fundamentals, principles and grammar
+
+## __Curriculum__
+
+* *Week 1* [What is python?](/week1)
+* *Week 2*
+* *Week 3*
+* *Week 4*
+* *Week 5*
+* *Week 6*
+* *Week 7*
+
+
+
