--- practice/hello.py
@@ -1,10 +1,0 @@
-#hello.py
-
-"""
-작성자 : 김문희
-작성일 : 2020-01-11
-기능: 화면출력
-"""
-
-print("Hello world")
-print("python")
